move the sembast rider dao
